dags/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.10.3/constraints-3.11.txt"
pydantic
urllib3
requests
python-dateutil
apache-airflow-providers-slack
oauth2client==4.1.3
requests_oauth2client==1.1.0
